Share ChildFirstClassLoader test utility via pinot-spi test-jar; cover skewed built-in copy
